Protegendo endpoints

+++ src/tasks/tasks.service.ts
@@ -4,6 +4,7 @@ import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { PaginationDto } from 'src/common/dto/pagination.dto';
+import { PayloadTokenDto } from 'src/auth/dto/payload-token.dto';
 
 @Injectable()
 export class TasksService {
@@ -36,14 +37,14 @@ export class TasksService {
     throw new HttpException('', HttpStatus.NOT_FOUND); //tratando o erro caso a task não existe
   }
 
-  async create(createTaskDto: CreateTaskDto) {
+  async create(createTaskDto: CreateTaskDto, tokenPayload: PayloadTokenDto) {
     try {
       const newTask = await this.prisma.task.create({
         data: {
           name: createTaskDto.name,
           description: createTaskDto.description,
           completed: false,
-          userId: createTaskDto.userId,
+          userId: tokenPayload.sub,
         },
       });
 
@@ -56,7 +57,11 @@ export class TasksService {
     }
   }
 
-  async update(id: number, updateTaskDto: UpdateTaskDto) {
+  async update(
+    id: number,
+    updateTaskDto: UpdateTaskDto,
+    tokenPayload: PayloadTokenDto,
+  ) {
     try {
       const findTask = await this.prisma.task.findFirst({
         where: {
@@ -70,6 +75,10 @@ export class TasksService {
           'Essa tarefa não existe!',
           HttpStatus.NOT_FOUND,
         );
+      }
+
+      if (findTask.userId !== tokenPayload.sub) {
+        throw new HttpException('Essa tarefa não é sua', HttpStatus.NOT_FOUND);
       }
 
       const task = await this.prisma.task.update({
@@ -96,7 +105,7 @@ export class TasksService {
     }
   }
 
-  async delete(id: number) {
+  async delete(id: number, tokenPayload: PayloadTokenDto) {
     try {
       const findTask = await this.prisma.task.findFirst({
         where: {
@@ -110,6 +119,10 @@ export class TasksService {
           'Essa tarefa não existe!',
           HttpStatus.NOT_FOUND,
         );
+      }
+
+      if (findTask.userId !== tokenPayload.sub) {
+        throw new HttpException('Essa tarefa não é sua', HttpStatus.NOT_FOUND);
       }
 
       await this.prisma.task.delete({
